optimization: use sequence for local Lampart clock Since Hibernate JPA 2 implementation still does not support scalar queries, I falled back to Hibernate Native SQL Query

+++ source/java/net/milanaleksic/mcs/domain/model/ModificationRepository.java
@@ -11,5 +11,5 @@ public interface ModificationRepository {
 
     void addModificationLog(ModificationType modificationType, String entityName, int id, String fieldName, Object fieldValue, int currentDatabaseVersion, long clock);
 
-    long getNextClock();
+    int getNextClock();
 }
